requirements/base.txt
pytz==2018.9  # https://github.com/stub42/pytz
python-slugify==2.0.1  # https://github.com/un33k/python-slugify
Pillow==5.4.1  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi

# Django
# ------------------------------------------------------------------------------
django==2.1.7  # pyup: < 2.1  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.1.2  # https://github.com/jazzband/django-model-utils
django-allauth==0.38.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.7.2  # https://github.com/django-crispy-forms/django-crispy-forms

# Django REST Framework
djangorestframework==3.9.1  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
django-filter==2.1.0

django-json-widget==0.2.0
Jinja2==2.10
networkx==2.2
numpy
scipy